LPS-51413 Deprecate BasicAuthSolrServer and retain it for backward compatibility
